clean up and consolidate end buttons
