+++ context/app/static/js/components/savedLists/SavedEntitiesTable/style.ts
@@ -1,8 +1,6 @@
 import { styled } from '@mui/material/styles';
 import Button from '@mui/material/Button';
-
 import { WhiteBackgroundIconButton } from 'js/shared-styles/buttons';
-import NumSelectedHeader from 'js/shared-styles/tables/NumSelectedHeader';
 
 const LeftMarginButton = styled(Button)(({ theme }) => ({
   marginLeft: theme.spacing(1),
@@ -12,10 +10,4 @@ const LeftMarginIconButton = styled(WhiteBackgroundIconButton)(({ theme }) => ({
   marginLeft: theme.spacing(1),
 }));
 
-const StyledNumSelectedHeader = styled(NumSelectedHeader)({
-  borderTop: 'none',
-  borderLeft: 'none',
-  borderRight: 'none',
-});
-
-export { LeftMarginButton, LeftMarginIconButton, StyledNumSelectedHeader };
+export { LeftMarginButton, LeftMarginIconButton };
