+++ context/app/static/js/components/workspaces/style.ts
@@ -5,6 +5,7 @@ import Description from 'js/shared-styles/sections/Description';
 import { WhiteBackgroundIconButton } from 'js/shared-styles/buttons';
 import TextField from '@mui/material/TextField';
 import SearchRoundedIcon from '@mui/icons-material/SearchRounded';
+import Button from '@mui/material/Button';
 
 const StyledButton = styled(WhiteBackgroundIconButton)(({ theme }) => ({
   height: theme.spacing(4.5),
@@ -47,6 +48,12 @@ const StyledSearchIcon = styled(SearchRoundedIcon)(({ theme }) => ({
   marginLeft: theme.spacing(1),
 }));
 
+const StyledLaunchButton = styled(Button)(({ theme }) => ({
+  border: `1px solid ${theme.palette.grey[300]}`,
+  borderRadius: theme.spacing(0.5),
+  padding: theme.spacing(1, 3),
+}));
+
 export {
   StyledButton,
   LinkButton,
@@ -56,4 +63,5 @@ export {
   StyledSubtitle2,
   StyledTextField,
   StyledSearchIcon,
+  StyledLaunchButton,
 };
