+++ context/app/static/js/shared-styles/tables/NumSelectedHeader/style.ts
@@ -3,7 +3,7 @@ import { styled } from '@mui/material/styles';
 const HeaderWrapper = styled('div')(({ theme }) => ({
   backgroundColor: theme.palette.white.main,
   zIndex: theme.zIndex.fileBrowserHeader,
-  border: `1px solid ${theme.palette.grey[300]}`,
+  borderBottom: `1px solid ${theme.palette.grey[300]}`,
   padding: theme.spacing(0.75, 2),
   gap: theme.spacing(1),
   width: '100%',
